qsieve linalg
