+++ src/qsieve/linalg.c
@@ -10,17 +10,15 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
-#include "fmpz.h"
-#include "qsieve.h"
-
 #ifdef __GNUC__
 # define memset __builtin_memset
 #else
 # include <string.h>
 #endif
-/*
-    Initialise linear algebra.
-*/
+
+#include "fmpz.h"
+#include "qsieve.h"
+
 void qsieve_linalg_init(qs_t qs_inf)
 {
     slong i, num_primes;
@@ -114,4 +112,42 @@ void qsieve_linalg_realloc(qs_t qs_inf)
     qs_inf->num_cycles = 0;
 
     memset(qs_inf->hash_table, 0, (1 << 20)*sizeof(mp_limb_t));
+}
+
+void qsieve_linalg_clear(qs_t qs_inf)
+{
+    slong i;
+
+    flint_free(qs_inf->relation);
+    flint_free(qs_inf->hash_table);
+    flint_free(qs_inf->table);
+
+    if (qs_inf->matrix != NULL)
+    {
+        for (i = 0; i < qs_inf->buffer_size; i++)
+        {
+            la_col_t * col = qs_inf->matrix + i;
+
+            if (col->weight)
+                flint_free(col->data);
+        }
+
+        flint_free(qs_inf->matrix);
+    }
+
+    if (qs_inf->Y_arr != NULL)
+    {
+        for (i = 0; i < qs_inf->buffer_size; i++)
+            fmpz_clear(qs_inf->Y_arr + i);
+        flint_free(qs_inf->Y_arr);
+    }
+
+    flint_free(qs_inf->prime_count);
+
+    qs_inf->relation = NULL;
+    qs_inf->matrix = NULL;
+    qs_inf->Y_arr = NULL;
+    qs_inf->prime_count = NULL;
+    qs_inf->hash_table = NULL;
+    qs_inf->table = NULL;
 }
